+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Add in sf plugins dependencies
   - Toggle merge driver from status bar
 - Fix startup error message
+- Add capability to debug internal sfdx-hardis commands by setting `vsCodeSfdxHardis.debugSfdxHardisCommands` to true
 
 ## [6.22.1] 2025-12-27